crewai-tutorials/1-Podcast_prepper/podcast_prepper/pyproject.toml
[tool.poetry]
name = "podcast_prepper"
version = "0.1.0"
description = "CrewAI-powered application designed for podcast hosts, helping them research guests, prepare detailed insights, and suggest relevant questions for upcoming episodes."
authors = ["Rok Benko"]


[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
python-dotenv = "*"
asyncio = "*"
crewai = { extras = ["tools"], version = ">=0.70.1,<1.0.0" }
exa-py = "*"

[tool.poetry.scripts]
start = "podcast_prepper.main:main"
plot = "podcast_prepper.main:plot"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
